+++ kipart/common.py
@@ -115,9 +115,9 @@ def issue(msg, level='warning'):
     
 def fix_pin_data(pin_data, part_num):
     '''Fix common errors in pin data.'''
-    fixed_pin_data = pin_data
-    if re.search('\s', pin_data) is not None:
-        fixed_pin_data = re.sub('\s', '_', pin_data)
+    fixed_pin_data = pin_data.strip()  # Remove leading/trailing spaces.
+    if re.search('\s', fixed_pin_data) is not None:
+        fixed_pin_data = re.sub('\s', '_', fixed_pin_data)
         issue("Replaced whitespace with '_' in pin '{pin_data}' of part {part_num}.".format(**locals()))
     return fixed_pin_data